+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @solid-design-system/docs
 
+## 2.18.0
+
+### Minor Changes
+
+- Added `sd-progress-bar`. _[`#3065`](https://github.com/solid-design-system/solid/pull/3065) [`e8de4cc`](https://github.com/solid-design-system/solid/commit/e8de4cccc9b405984cba761233c3d878bb3ea154) [@auroraVasconcelos](https://github.com/auroraVasconcelos)_
+
 ## 2.17.2
 
 ### Patch Changes