+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @solid-design-system/tokens
 
+## 6.23.0
+
+### Minor Changes
+
+- Add `progress-bar`new tokens: _[`#3065`](https://github.com/solid-design-system/solid/pull/3065) [`e8de4cc`](https://github.com/solid-design-system/solid/commit/e8de4cccc9b405984cba761233c3d878bb3ea154) [@auroraVasconcelos](https://github.com/auroraVasconcelos)_
+  - --sd-progress-bar\_\_slide-bar--inverted-color-background
+  - --sd-progress-bar\_\_slide-bar-color-background
+  - --sd-progress-bar--active--inverted-color-background
+
 ## 6.22.2
 
 ## 6.22.1